# swagger-impl/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Channel Reservation",
    "description": "Using NodeJS, Swagger, RabbitMQ and Websockets, the purpose of this seed project is to build an elastic infra structure that can async message between clients in a non stateful manner.",
    "version": "1.0.0"
  },
  "produces": [
    "application/json"
  ],
  "host": "localhost:8080",
  "basePath": "/v1",
  "paths": {
    "/channels/{id}": {
      "put": {
        "summary": "modify a channel",
        "description": "Used to create channel.",
        "x-swagger-router-controller": "Default",
        "tags": [
          "Default"
        ],
        "operationId": "channelsIdPut",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Channel id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Field to change.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ModifyChannel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucess"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/channels": {
      "get": {
        "summary": "Channels",
        "description": "Used to find all channels.",
        "x-swagger-router-controller": "Default",
        "tags": [
          "Default"
        ],
        "operationId": "channelsGet",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "description": "The mode of the channel open or closed",
            "required": true,
            "type": "string",
            "enum": [
              "open",
              "closed"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "An array of channels that match the mode",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Channel"
              }
            }
          },
          "default": {
            "description": "Message",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      },
      "post": {
        "summary": "Create Channel",
        "description": "Used to create channel.",
        "x-swagger-router-controller": "Default",
        "tags": [
          "Default"
        ],
        "operationId": "channelsPost",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "basic create channel data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreatChannel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of channels that match the mode"
          },
          "default": {
            "description": "Message",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Channel": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Unique identifier representing a specific channel"
        },
        "desc": {
          "type": "string",
          "description": "A human readable description of channel."
        },
        "href": {
          "type": "string",
          "description": "a link to the channel."
        },
        "rel": {
          "type": "string",
          "description": "The rel attribute specifies the relationship between the current document and the linked document.",
          "default": "subscribe"
        }
      }
    },
    "ModifyChannel": {
      "type": "object",
      "properties": {
        "desc": {
          "type": "string",
          "description": "A human readable description of channel."
        },
        "mode": {
          "type": "string",
          "description": "The mode of the channel open or closed",
          "enum": [
            "open",
            "closed"
          ]
        }
      }
    },
    "CreatChannel": {
      "type": "object",
      "properties": {
        "desc": {
          "type": "string",
          "description": "A human readable description of channel."
        }
      }
    },
    "Message": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    }
  }
}